rename package to servicemonitor
